+++ src/cpp/session/include/session/SessionSourceDatabase.hpp
@@ -220,6 +220,7 @@ struct Events : boost::noncopyable
 {
    RSTUDIO_BOOST_SIGNAL<void(boost::shared_ptr<SourceDocument>)>                     onDocUpdated;
    RSTUDIO_BOOST_SIGNAL<void(const std::string&, boost::shared_ptr<SourceDocument>)> onDocRenamed;
+   RSTUDIO_BOOST_SIGNAL<void(boost::shared_ptr<SourceDocument>)>                     onDocReopened;
    RSTUDIO_BOOST_SIGNAL<void(boost::shared_ptr<SourceDocument>)>                     onDocAdded;
    RSTUDIO_BOOST_SIGNAL<void(boost::shared_ptr<SourceDocument>)>                     onDocPendingRemove;
    RSTUDIO_BOOST_SIGNAL<void(const std::string&, const std::string&)>                onDocRemoved;