+++ NEWS.md
@@ -54,7 +54,7 @@
 - (#rstudio-pro/8846): Improved contrast of check boxes on Workbench homepage to meet 3:1 minimum contrast (AA)
 
 ### Dependencies
-- Copilot Language Server 1.381.0
+- Copilot Language Server 1.387.0
 - Electron 38.3.0
 - Node.js 22.18.0
 - Quarto 1.8.25